Fix for TuneIn's broken media receiver
When you ask the TuneIn receiver to pause/play playback with a correct request Id, it'll answer with request id 0, regardless of the senders request id.
Furthermore it constantly sends device playback updates although nothing has changed! That is annoying but other than spamming the logs and network it doesn't do much harm.